barter_system/requirements.txt
django
djangorestframework
psycopg2-binary
requests
django-environ
djangorestframework-simplejwt==5.2.2
pytest
pytest-django
